+++ agir/front/components/app/utils.js
@@ -1,5 +1,5 @@
 import React, { lazy as reactLazy, useEffect, useMemo, useState } from "react";
-import { useIsOffline } from "@agir/front/allPages/offline/hooks";
+import { useIsOffline } from "@agir/front/offline/hooks";
 
 export const lazy = (lazyImport, fallback) => {
   const LazyComponent = (props) => {
